+++ src/op3/__init__.py
@@ -1,3 +1,5 @@
 """op3 package."""
 
-__version__ = "0.1.14"
+from opstree._version import __version__
+
+__all__ = ["__version__"]